backend/requirements.txt
pip
fastapi
python-dotenv
uvicorn
requests
pymongo
passlib
bcrypt==4.0.1
python-jose[cryptography]